[Feat] #71- 설정 뷰 서비스 의견 제안 외부 링크 연결
